tests/wpt/tests/svg/animations/repeat-iteration-event-001.svg

<svg xmlns="http://www.w3.org/2000/svg">
  <title>Animation does not triggers on 'repeat(0)'</title>
  <script xmlns="http://www.w3.org/1999/xhtml" src="/resources/testharness.js"/>
  <script xmlns="http://www.w3.org/1999/xhtml" src="/resources/testharnessreport.js"/>
  <rect width="100" height="100" fill="green">
    <animate id="anim" attributeName="visibility" to="visible" begin="0s" dur="2s" repeatCount="4"/>
    <set attributeName="width" to="0" begin="anim.repeat(0)"/>
  </rect>
  <script>
  function waitForFrame() {
    return new Promise(resolve => {
      window.requestAnimationFrame(resolve);
    });
  }

  promise_test(t => {
    let set = document.querySelector("set");
    new EventWatcher(t, set, 'beginEvent');
    let windowWatcher = new EventWatcher(t, window, 'load');
    return windowWatcher.wait_for('load').then(() => {
      return waitForFrame();
    }).then(() => {
      assert_equals(set.parentNode.getBBox().width, 100);
    });
  });
  </script>
</svg>
